Allow owners to edit/delete their claims

+++ api.py
@@ -129,7 +129,7 @@ def fragment(fn):
         ret = fn(*args, **kwargs)
         if isinstance(ret, CoroutineType):
             ret = await ret
-        elif isinstance(ret, HTTPResponse):
+        if isinstance(ret, HTTPResponse):
             return ret
         return html(ret)
 
@@ -673,10 +673,14 @@ async def search(request):
 
 
 @app.get("/claims/<claim_id>/edit")
-@admin_only  # until claims have an owner
 @fragment
 async def edit_claim_form(request, claim_id: int):
     claim = O.Claim(claim_id)
+    if claim.owner.id != context.user.id:
+        return HTTPResponse(
+            body="403 Forbidden",
+            status=403,
+        )
     return f"""
         <form
             action="/claims/{claim_id}/edit"
@@ -689,10 +693,14 @@ async def edit_claim_form(request, claim_id: int):
 
 
 @app.delete("/claims/<claim_id>")
-@admin_only
 @fragment
 async def delete_claim(request, claim_id: int):
     claim = O.Claim(claim_id)
+    if claim.owner.id != context.user.id:
+        return HTTPResponse(
+            body="403 Forbidden",
+            status=403,
+        )
     claim.delete()
     return """
         <meta http-equiv="refresh" content="0; url=/">
@@ -700,13 +708,17 @@ async def delete_claim(request, claim_id: int):
 
 
 @app.post("/claims/<claim_id>/edit")
-@admin_only
 async def edit_claim(request, claim_id: int):
     form = D(request.form)
     if "value" in request.files:
         f = request.files["value"][0]
         form["value"] = f"data:{f.type};base64,{base64.b64encode(f.body).decode()}"
     claim = O.Claim(claim_id)
+    if claim.owner.id != context.user.id:
+        return HTTPResponse(
+            body="403 Forbidden",
+            status=403,
+        )
     value = form.get("value")
     if claim.verb.data_type.name.endswith("directed_link"):
         # no longer allowed
